requirements.txt
django==5.2.*
django-debug-toolbar==5.2.*
Pillow==11.2.*
environs[django]==14.2.*
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
requests==2.32.5